Fixing TZ issue
