+++ src/test/resources/application-test.yml
@@ -17,3 +17,6 @@ auth:
   generic-user:
     username: Tyler
     password: Changeme!
+
+server:
+  timezone: America/Chicago
